r4941/sofa-dev : FIX: other warnings on windows.
 //partial sync (4/6 changes)//

+++ modules/sofa/component/linearsolver/BTDLinearSolver.h
@@ -891,7 +891,7 @@ public:
             if (j0 == n)
             {
                 // i0 row is the identity
-                Inv.set(i0,i0,1.0/M.element(i0,i0));
+                Inv.set(i0,i0,(float)1.0/M.element(i0,i0));
                 ++i0;
             }
             else break;
